# docs/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gnovium Knowledge OS API",
    "version": "1.0.1",
    "description": "Gnovium API is a local-first Knowledge Operating System API for building applications where knowledge behaves like a living system rather than a collection of disconnected documents.\n\nThis API supports:\n- Block-based content management with rich text editing\n- Structured relational knowledge modeling with typed entities and properties\n- Interactive knowledge graph with traversal, pathfinding, and queries\n- Git-inspired versioning with branches, snapshots, and visual diffs\n- AI-powered semantic search and natural language question answering\n- Workspace governance with health scoring and duplicate detection\n- Cross-device synchronization with conflict resolution\n- File management with presigned URL uploads (cloud mode)\n\nDeployment modes:\n- **Local Mode** (default): SQLite storage, offline-first, local AI via Inference Runtime\n- **Cloud Mode**: PostgreSQL, S3 storage, managed infrastructure",
    "contact": {
      "name": "Gaurav Kaloliya — Founder & Creator",
      "url": "https://www.linkedin.com/in/gaurav-kaloliya-b44569417",
      "email": "gaurav@gnovium.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/GauravKaloliya/gnovium/blob/main/LICENSE"
    },
    "termsOfService": "https://gnovium.com/terms"
  },
  "servers": [
    {
      "url": "https://api.gnovium.com/api/v1",
...
    },
    {
...
    },
    {
...
    }
  ],
  "paths": {
...
    "/tags/": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "List tags",
        "description": "Returns all tags for a workspace.",
        "operationId": "tagsList",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace scope for the operation."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Page number. Defaults to 1."
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Page size from 1 to 100. Defaults to 25."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Create tag",
        "description": "Creates a new tag with an optional color. Tags are lightweight labels that can be applied to entities.",
        "operationId": "tagsCreate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/tags/{tag_id}/entities/{entity_id}": {
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Tag an entity",
        "description": "Applies a tag to an entity. Both the tag and entity must exist in the same workspace.",
        "operationId": "tagsTagEntity",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tag ID"
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Entity ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "summary": "Untag an entity",
        "description": "Removes a tag from an entity. The tag itself is not deleted.",
        "operationId": "tagsUntagEntity",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tag ID"
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Entity ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/search/": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search workspace",
        "description": "Searches entities and blocks by keyword, PostgreSQL full text, hybrid, or semantic mode. Hybrid mode combines keyword matching with embedding similarity for the best results.",
        "operationId": "searchQuery",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace scope for the operation."
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query. Minimum 1 character."
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search mode: keyword, full_text, hybrid (default), or semantic."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Maximum results. Range 1–50. Defaults to 20."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via OAuth2 or direct login. Format: Bearer <token>"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for server-to-server authentication. Generate from workspace settings."
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.gnovium.com/oauth/authorize",
            "tokenUrl": "https://api.gnovium.com/oauth/token",
            "scopes": {
              "workspace:read": "Read workspace metadata and content",
              "workspace:write": "Create, update, and delete workspace content",
              "workspace:admin": "Manage workspace settings, members, and billing",
              "user:read": "Read user profile information",
              "user:write": "Update user profile settings",
              "offline_access": "Receive refresh tokens for long-lived access"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
      "DataEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  },
...
}